Removal of addon. Added loop detection and sync
The addon would be hard to use so it's best to use the supported audio nodes. They even support custom effects like reverb!

The project now detects which loop it's in (and it's pretty accurate) so that polos will be able to load the correct loop sound.

The progress bar could sometimes desync, so it's been fixed so it teleports itself to the correct position if it falls behind the actual current loop.