More optimizations
The sound delay thing was partly my fault as it would actually play without a noticeable delay. Let's just say I was using a sound with a delay in the actual file.

Still, I optimized the code even more so that the thingy knows where is your mouse without the need of collision boxes, which can be unreliable when the mouse moves fast.

The tool will now crash if not enough sounds are present. For testing you can comment lines 9, 10 and 11 in `res://Scripts/AudioPlayer.gd`