Fixed a bug that would mess with loops + feature
There was a bug that would trigger at a specific frame in the transition between loops. The sound would still play when the polo is unpicked, and the loop would be incorrect. It's been fixed now, I think.

Also, there's a new feature that now makes the polos have their signature random heights!